requirements.txt

Pillow==8.3.1
requests==2.26.0
celery==5.4.0
Django==1.9.9
django-environ==0.12.0
django-redis==5.4.0
pathspec==0.12.1
redis==5.2.1

